Fix an issue when a user's password contains <,> or & characters
Issue: https://github.com/wso2/product-apim/issues/8153
This fixes the issue "Error: stxUnexpectedCharException:
Unexpected character when login to the admin console". In here we need
to replace XML reserved characters with the relevant character entities.
"&" should be replaced with "&amp;" and "<" should be replaced with "&lt;"
and ">" should be replaced with "&gt;".